+++ src/overlays/ovl_80_shared_board/EAB00.h
@@ -6,7 +6,7 @@
 
 s32 func_800D7250_EAE70_shared_board(UnkEA790Struct* arg0, s32 arg1);
 s32 func_800D742C_EB04C_shared_board(UnkEA790Struct* arg0, s16 playerNo, s32 arg2);
-s16 func_800D7518_EB138_shared_board(UnkEA790Struct* arg0);
+s32 func_800D7518_EB138_shared_board(UnkEA790Struct* arg0);
 void func_800D7568_EB188_shared_board(UnkBoard2* arg0, Vec* arg1, Vec* arg2, f32 arg3);
 UnkEA790Struct* func_800D76A0_EB2C0_shared_board(s16 arg0, s16* arg1);
